teeetet

+++ src/user/user.service.ts
@@ -1,19 +1,49 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
-import { User } from './entities/user.entity';
+import { User } from './entity/user.entity';
 import { Repository } from 'typeorm';
+import * as bcrypt from 'bcrypt';
+import { authVariableKeys } from '../common/const/auth.const';
+import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(User)
-    private userRepository: Repository<User>
+    private readonly userRepository: Repository<User>,
+    private readonly configService: ConfigService
   ) {
   }
-  create(createUserDto: CreateUserDto) {
-    return this.userRepository.save(createUserDto);
+  async create(createUserDto: CreateUserDto) {
+    const { email, password } = createUserDto;
+
+    const checkEmail = await this.userRepository.exists({
+      where: {
+        email,
+      },
+    });
+
+    if (checkEmail) {
+      throw new BadRequestException('이미 존재하는 이메일입니다!');
+    }
+
+    const hash = await bcrypt.hash(
+      password,
+      this.configService.get<number>(authVariableKeys.hashRounds),
+    );
+
+    await this.userRepository.save({
+      email,
+      password: hash,
+    });
+
+    return this.userRepository.findOne({
+      where: {
+        email,
+      },
+    });
   }
 
   findAll() {
@@ -35,6 +65,8 @@ export class UserService {
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const { password } = updateUserDto;
+    
     const user = await this.userRepository.findOne({
       where: {
         id
@@ -45,10 +77,17 @@ export class UserService {
       throw new NotFoundException('존재하지 않는 사용자 입니다.');
     }
 
+    const hashedPassword = await bcrypt.hash(
+      password, 
+      this.configService.get<number>(authVariableKeys.hashRounds)
+    );
+    
     await this.userRepository.update(
       { id },
-      updateUserDto
-    );
+      {
+        ...updateUserDto,
+        password: hashedPassword
+      });
 
     return this.userRepository.findOne({
       where: {
@@ -57,7 +96,19 @@ export class UserService {
     });
   }
 
-  remove(id: number) {
-    return this.userRepository.delete(id);
+  async remove(id: number) {
+    const user = await this.userRepository.findOne({
+      where: {
+        id
+      }
+    });
+
+    if (!user) {
+      throw new NotFoundException('존재하지 않는 사용자 입니다.');
+    }
+
+    await this.userRepository.delete(id);
+
+    return id;
   }
 }
